Convert stp-solve-base.rkt to typed/racket
Finished conversion of solve-base.  Most significant change is that the *ms-array* array
for looking up better-move-schemas (used when building the even-better-MS space-index) now
indexes piece-types at one-less than the piece-type.  That is, starting with 1, piece-type i
is found in the table at i-1.

During this same commit, I moved the definition and creation of *expansion-space* from stp-init.rkt
to stp-spaceindex.rkt.  This was motivated to reduce amount of cross-boundary processing between
untyped and typed modules and objects.  That helped greatly but it is still slower than we started out.