Re-segmenting Fringes (started)
Minor refactoring, moving *most-pos/neg-fixnum* to stp-init and compute-segment-bounds to stp-fringefilerep.
Also added guard on creating the fringehead in remove-dupes so that no files will even be opened if
*late-duplicate-removal* is true.

Started on resegment-fringe in stp-fringefilerep but much yet to do there.  Need to loop through the
fringehead by advance-fhead!, but then switching output files when the current position is no longer
less than the currently relevant segment-bound.