Proto-fringe distribution by offset
Each worker sends its proto-fringe segments to the other workers in semi-coordinated
order.  Each worker first sends the file destined for next-higher worker.  Then to the
worker two above it, then three, and so on.  Modulo arithmetic is used to wrap around
the number of workers.  The hope is that this will reduce network contention using
the switch more efficiently.